pyproject.toml
[project]
name = "mcp_playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.4.9.2",
    "autogen-ext[grpc,mcp,ollama,openai]>=0.4.9.2",
    "openai-agents>=0.0.15",
    "mcp[cli]>=1.5.0",
    "mcp-server-fetch>=2025.1.17",
    "python-dotenv>=1.0.1",
    "ipython>=9.4.0",
]
